Bugfix amt set and get did not handle hash collision
Upon hash bits exhaustion we use the bits from the key.

While this is not an optimal source of bits due to the non-random nature
of the ascii characters in the keys which results in non optimal use of
RAM, it should be good enough for use here under the assumption that
hash collisions are rare.